ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
paddleocr==2.7.0.3
paddlepaddle==2.5.2
numpy==1.26.2
Pillow==10.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pydantic==2.5.2
